splitting unpack_ac from ac to use in other modules
